define scores helpers
